modularizando navbar
