+++ client/src/components/Search.jsx
@@ -1,14 +1,7 @@
 import * as React from 'react';
 import { styled, } from '@mui/material/styles';
-import AppBar from '@mui/material/AppBar';
-import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
-
-
-import InputBase from '@mui/material/InputBase';
-
+import { Autocomplete, IconButton, TextField, Toolbar } from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
-import { Autocomplete, capitalize, CssBaseline, IconButton, TextField } from '@mui/material';
 import { useDispatch, useSelector } from 'react-redux';
 import { useState } from 'react';
 import { generalSearch } from '../actions/search';
@@ -19,7 +12,7 @@ const Search = styled('div')(({ theme }) => ({
   backgroundColor:'white',
   opacity: '90%',
   borderRadius: theme.shape.borderRadius,
-  boxShadow: "0px 3px 10px -2px rgb(0 0 0 / 30%)",
+  boxshadow: "0px 3px 10px -2px rgb(0 0 0 / 30%)",
   '&:hover': {
     backgroundColor: 'white',
     opacity:'100%'
@@ -33,8 +26,6 @@ const Search = styled('div')(({ theme }) => ({
     width: '500px',
   },
 }));
-
-
 
 
 export const SearchBar = ()=> {
@@ -52,59 +43,53 @@ export const SearchBar = ()=> {
       
     }
     const handleClickButton = (e) =>{
-      e.preventDefault();
+      // e.preventDefault();
       if(search.length > 1)dispatch(generalSearch(search))
 
     }
 
   return (
-
-
-      // <AppBar position="relative" sx={{top:{xs:55, md:68}, height:'4rem', alignItems:'center', justifyContent:'center', bgcolor:'transparent'}}>
-        <Toolbar sx={{ alingItems:'center', bgcolor:'white', boxShadow:'0 0 0 0 ', display:'flex', justifyContent:'end', alingItems:'center', bgcolor:'transparent', my:-1.7}} >
-          <Search  >
-             <Autocomplete
-                freeSolo
-                variant=''
-                id="searchBar-ecomerce"
-                disableClearable
-                value={search}
-                options={allProducts.map( el => `${el.nombre} de -> ${el.marcaId.nombre} #${el.id}`.toLocaleUpperCase())}
-                onInputChange={hadleInputChange}
-                onChange={handleChangeAutoComplete}
+    <Toolbar sx={{display:'flex', justifyContent:'end', alingItems:'center', bgcolor:'transparent', my:{xs:0, sm:0,  md:-1.7,} }} >
+      <Search >
+          <Autocomplete
+            freeSolo
+            variant=''
+            id="searchBar-ecomerce"
+            disableClearable
+            value={search}
+            options={allProducts.map( el => `${el.nombre} de -> ${el.marcaId.nombre} #${el.id}`.toLocaleUpperCase())}
+            onInputChange={hadleInputChange}
+            onChange={handleChangeAutoComplete}
+            renderInput={(params) => (
+              <TextField
                 onSubmit={handleClickButton}
+                placeholder='Buscar'
+                sx={{ 
+                  border: 'transparent', 
+                  fontSize:'0.5rem',
+                  '& input': {
+                    height:2,
+                    margin:0,
+                    padding:0,  
+                    color: (theme) =>
+                      theme.palette.getContrastText(theme.palette.background.paper),
+                  },
+                  }}
+                {...params}
+                InputProps={{
+                  ...params.InputProps,
+                  type: 'search',
+                  boxshadow:'0px 2px 4px -1px rgb(0 0 0 / 20%)',
 
-                renderInput={(params) => (
-                  <TextField
-                    placeholder='Buscar'
-                    sx={{ 
-                      border: 'transparent', 
-                      '& input': {
-                        height:2,
-                        margin:0,
-                        padding:0,  
-                        color: (theme) =>
-                          theme.palette.getContrastText(theme.palette.background.paper),
-                      },
-                     }}
-                    {...params}
-                    InputProps={{
-                      ...params.InputProps,
-                      type: 'search',
-                      boxShadow:'0px 2px 4px -1px rgb(0 0 0 / 20%)',
-
-                    }}
-                    
-                  />
-                )}
+                }}
+                
               />
-          </Search>
-          <IconButton onClick={handleClickButton} sx={{top:"3px"}} >
-            <SearchIcon  />
-          </IconButton>
-
-        </Toolbar>
-      // </AppBar>
-
-  );
+            )}
+          />
+      </Search>
+      <IconButton onClick={handleClickButton} sx={{top:"3px"}} >
+        <SearchIcon  />
+      </IconButton>
+    </Toolbar>
+  )
 }
