+++ client/src/helpers/userMenuLinks.js
@@ -1,0 +1,32 @@
+export const logged = [
+    {
+      label: 'Mi perfil',
+      link: '/usuario/nombre'
+    },
+    {
+      label: 'Lista de deseos',
+      link: '/usuario/nombre/lista-de-deseos'
+    },{
+      label: 'Compras',
+      link: '/usuario/nombre/compras'
+    },
+  ]
+  
+  export const admin =[
+    {
+      label: 'Mi perfil',
+      link: '/admin/nombre'
+    },
+    {
+      label: 'Lista de deseos',
+      link: '/admin/nombre/lista-de-deseos'
+    },
+    {
+      label: 'Dashboard',
+      link: '/admin/nombre/dashboard'
+    },
+    {
+      label: 'Administracion de productos',
+      link: '/admin/nombre/administracion'
+    },
+  ]
